third_party/blink/web_tests/platform/generic/fast/css/font-face-unicode-range-monospace-expected.txt

@font-face should not affect width of characters outside of its unicode-range

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('test1').offsetWidth is document.getElementById('ref').offsetWidth
PASS document.getElementById('test2').offsetWidth is document.getElementById('ref').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE

m
m
m
